requirements.txt
streamlit
numpy
pillow
tensorflow
opencv-python
joblib
scikit-learn
matplotlib
pydub
